backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
google-api-python-client==2.155.0
google-auth==2.37.0
httpx==0.27.2
apscheduler==3.10.4